requirements.txt
djangorestframework==3.9.4
Django==2.2.3

